Separate edit trade modal into a component
