+++ client/src/app/styles/dashboard/trade-history.css
@@ -211,90 +211,3 @@
     padding: 4px 6px;
   }
 }
-
-.modal-bg {
-  position: fixed;
-  inset: 0;
-  background: rgba(0,0,0,0.55);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 9999;
-}
-
-.modal-box {
-  background: #ffffff;
-  width: 420px;
-  padding: 28px;
-  border-radius: 12px;
-  box-shadow: 0 8px 30px rgba(0,0,0,0.25);
-  animation: modalPop 0.25s ease-out;
-}
-
-@keyframes modalPop {
-  from {
-    transform: scale(0.85);
-    opacity: 0;
-  }
-  to {
-    transform: scale(1);
-    opacity: 1;
-  }
-}
-
-.modal-box h3 {
-  font-size: 20px;
-  margin-bottom: 14px;
-  font-weight: 600;
-}
-
-.modal-box label {
-  display: block;
-  margin: 12px 0 6px;
-  font-size: 14px;
-  color: #444;
-}
-
-.modal-box input,
-.modal-box select {
-  width: 100%;
-  padding: 10px 12px;
-  border-radius: 6px;
-  border: 1px solid #ccc;
-  outline: none;
-  font-size: 14px;
-}
-
-.modal-box input:focus,
-.modal-box select:focus {
-  border-color: #2563eb;
-}
-
-.btn-row {
-  display: flex;
-  justify-content: flex-end;
-  margin-top: 20px;
-  gap: 10px;
-}
-
-.cancel-btn,
-.save-btn {
-  padding: 10px 18px;
-  border-radius: 6px;
-  border: none;
-  cursor: pointer;
-  font-size: 14px;
-}
-
-.cancel-btn {
-  background: #e5e5e5;
-}
-
-.save-btn {
-  background: #2563eb;
-  color: #fff;
-}
-
-.save-btn:hover {
-  background: #1e4fc7;
-}
